fix(services): mask comments and strings before the project scan records a declaration
ProjectPathScanner skipped a line only when it started with `#` or `//`, so a
declaration inside `<# ... #>`, inside a `<#>` marker's body or after a mid-line
`#` was recorded in the project declaration cache as live code - and a
commented-out module was pushed onto the module stack, prefixing the fullPath of
every real declaration below it.

moduleDeclarations already solved this for its own scan. Its
maskCommentsAndStrings and indentOf move to src/utils/verseText, and the scanner
splits the masked text instead of growing a third encoding of the three Verse
comment rules.
