+++ changelog.d/292.fixed.md
@@ -1,0 +1,1 @@
+**Project scan**: A declaration commented out with a block comment, a `<#>` marker or a mid-line `#` is no longer recorded in the project declaration cache as live code.
